+++ monailabel/endpoints/proxy.py
@@ -60,6 +60,8 @@ async def proxy_dicom(op: str, path: str, response: Response):
             else settings.MONAI_LABEL_QIDO_PREFIX
             if op == "qido"
             else settings.MONAI_LABEL_STOW_PREFIX
+            if op == "stow"
+            else ""
         )
 
         # some version of ohif requests metadata using qido so change it to wado
@@ -72,7 +74,10 @@ async def proxy_dicom(op: str, path: str, response: Response):
             proxy_path = f"{server}/{path}"
 
         logger.debug(f"Proxy connecting to /dicom/{op}/{path} => {proxy_path}")
-        timeout = httpx.Timeout(5.0, read=settings.MONAI_LABEL_DICOMWEB_READ_TIMEOUT)
+        timeout = httpx.Timeout(
+            settings.MONAI_LABEL_DICOMWEB_PROXY_TIMEOUT,
+            read=settings.MONAI_LABEL_DICOMWEB_READ_TIMEOUT,
+        )
         proxy = await client.get(proxy_path, timeout=timeout)
 
     response.body = proxy.content
@@ -93,3 +98,9 @@ async def proxy_qido(path: str, response: Response, user: User = Depends(get_bas
 @router.get("/dicom/stow/{path:path}", include_in_schema=False)
 async def proxy_stow(path: str, response: Response, user: User = Depends(get_basic_user)):
     return await proxy_dicom("stow", path, response)
+
+
+# https://fastapi.tiangolo.com/tutorial/path-params/#order-matters
+@router.get("/dicom/{path:path}", include_in_schema=False)
+async def proxy(path: str, response: Response, user: User = Depends(get_basic_user)):
+    return await proxy_dicom("", path, response)