+++ plugins/qupath/src/main/java/qupath/lib/extension/monailabel/commands/SubmitLabel.java
@@ -16,8 +16,8 @@ package qupath.lib.extension.monailabel.commands;
 import java.awt.image.BufferedImage;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 
@@ -37,11 +37,17 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 
 import qupath.lib.extension.monailabel.MonaiLabelClient;
+import qupath.lib.extension.monailabel.MonaiLabelClient.ImageInfo;
 import qupath.lib.extension.monailabel.Utils;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.images.ImageData;
+import qupath.lib.images.writers.ImageWriterTools;
 import qupath.lib.objects.PathObject;
+import qupath.lib.plugins.parameters.ParameterList;
+import qupath.lib.regions.RegionRequest;
+import qupath.lib.roi.ROIs;
+import qupath.lib.roi.interfaces.ROI;
 
 public class SubmitLabel implements Runnable {
 	private final static Logger logger = LoggerFactory.getLogger(SubmitLabel.class);
@@ -54,37 +60,74 @@ public class SubmitLabel implements Runnable {
 
 	@Override
 	public void run() {
-		Path path = null;
+		Path annotationXML = null;
+		Path imagePatch = null;
 		try {
 			var viewer = qupath.getViewer();
 			var imageData = viewer.getImageData();
 			String image = Utils.getNameWithoutExtension(imageData.getServerPath());
-			path = getAnnotationsXml(image, imageData);
-			logger.info("Annotations XML: " + path);
 
-			String params = "{}";
+			if (!MonaiLabelClient.imageExists(image)) {
+				logger.info("Image does not exist on Server.");
+				var selected = imageData.getHierarchy().getSelectionModel().getSelectedObject();
+				var roi = selected != null ? selected.getROI() : null;
+				int[] bbox = Utils.getBBOX(roi);
+				if (bbox[2] <= 0 && bbox[3] <= 0) {
+					Dialogs.showWarningNotification("MONAILabel",
+							"Please select the Annotation ROI/Rectangle for submission");
+					return;
+				}
 
-			if (Dialogs.showYesNoDialog("MONAILabel", "Do you like to submit this annotation to MONAI Label Server?\n\n"
-					+ "This might override any existing annotations in MONAI Label server for: '" + image + "'")) {
-				String res = MonaiLabelClient.saveLabel(image, path.toFile(), null, params);
-				logger.info("SUBMIT:: resp = " + res);
-				Dialogs.showInfoNotification("MONALabel", "Label/Annotations saved in Server");
+				String patchName = image + String.format("-patch-%d_%d_%d_%d", bbox[0], bbox[1], bbox[2], bbox[3]);
+				ParameterList list = new ParameterList();
+				list.addStringParameter("Location", "Patch (x,y,w,h)", Arrays.toString(bbox));
+				list.addStringParameter("Patch", "Patch Name", patchName);
+
+				if (Dialogs.showParameterDialog("MONAILabel - Save Patch + Label", list)) {
+					bbox = Utils.parseStringArray(list.getStringParameterValue("Location"));
+					patchName = list.getStringParameterValue("Patch");
+					if (Dialogs.showYesNoDialog("MONAILabel",
+							"This will upload BOTH image patch + annotation to MONAI Label Server.\n\n"
+									+ "Do you want to continue?")) {
+
+						annotationXML = getAnnotationsXml(image, imageData, bbox);
+						logger.info("MONAILabel:: Annotations XML: " + annotationXML);
+
+						imagePatch = java.nio.file.Files.createTempFile("patch", ".png");
+						var requestROI = RegionRequest.createInstance(imageData.getServer().getPath(), 1, roi);
+						ImageWriterTools.writeImageRegion(imageData.getServer(), requestROI, imagePatch.toString());
+
+						ImageInfo imageInfo = MonaiLabelClient.saveImage(patchName, imagePatch.toFile(), "{}");
+						logger.info("MONAILabel:: New Image ID => " + imageInfo.image);
+						Dialogs.showInfoNotification("MONALabel", "Image Patch uploaded to MONAILabel Server");
+
+						MonaiLabelClient.saveLabel(imageInfo.image, annotationXML.toFile(), null, "{}");
+						Dialogs.showInfoNotification("MONALabel", "Label/Annotations saved in Server");
+					}
+				}
+			} else {
+				if (Dialogs.showYesNoDialog("MONAILabel",
+						"Do you like to submit this annotation to MONAI Label Server?\n\n"
+								+ "This might override any existing annotations in MONAI Label server for: '" + image
+								+ "'")) {
+
+					annotationXML = getAnnotationsXml(image, imageData, new int[4]);
+					logger.info("Annotations XML: " + annotationXML);
+
+					MonaiLabelClient.saveLabel(image, annotationXML.toFile(), null, "{}");
+					Dialogs.showInfoNotification("MONALabel", "Label/Annotations saved in Server");
+				}
 			}
 		} catch (Exception ex) {
 			ex.printStackTrace();
-			Dialogs.showErrorMessage("MONAI Label - Pathology", ex);
+			Dialogs.showErrorMessage("MONAILabel - Pathology", ex);
 		} finally {
-			if (path != null) {
-				try {
-					Files.deleteIfExists(path);
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
-			}
+			Utils.deleteFile(annotationXML);
+			Utils.deleteFile(imagePatch);
 		}
 	}
 
-	private Path getAnnotationsXml(String image, ImageData<BufferedImage> imageData)
+	private Path getAnnotationsXml(String image, ImageData<BufferedImage> imageData, int[] bbox)
 			throws IOException, ParserConfigurationException, TransformerException {
 		DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
 		DocumentBuilder docBuilder = docFactory.newDocumentBuilder();
@@ -97,24 +140,35 @@ public class SubmitLabel implements Runnable {
 		Element annotations = doc.createElement("Annotations");
 		annotations.setAttribute("Name", "");
 		annotations.setAttribute("Description", "");
-		annotations.setAttribute("X", "0");
-		annotations.setAttribute("Y", "0");
-		annotations.setAttribute("W", "0");
-		annotations.setAttribute("H", "0");
+		annotations.setAttribute("X", String.valueOf(bbox[0]));
+		annotations.setAttribute("Y", String.valueOf(bbox[1]));
+		annotations.setAttribute("W", String.valueOf(bbox[2]));
+		annotations.setAttribute("H", String.valueOf(bbox[3]));
 		rootElement.appendChild(annotations);
+
+		ROI patchROI = (bbox[2] > 0 && bbox[3] > 0) ? ROIs.createRectangleROI(bbox[0], bbox[1], bbox[2], bbox[3], null)
+				: null;
 
 		int count = 0;
 		var groups = new HashMap<String, String>();
 		List<PathObject> objs = imageData.getHierarchy().getFlattenedObjectList(null);
 		for (int i = 0; i < objs.size(); i++) {
 			var a = objs.get(i);
+
+			// Ignore which doesn't have class
 			String name = a.getPathClass() != null ? a.getPathClass().getName() : null;
 			if (name == null || name.isEmpty()) {
 				continue;
 			}
 
+			// Ignore Points
 			var roi = a.getROI();
 			if (roi == null || roi.isPoint()) {
+				continue;
+			}
+
+			// Ignore other objects not part of BBOX
+			if (patchROI != null && !patchROI.contains(roi.getCentroidX(), roi.getCentroidY())) {
 				continue;
 			}
 
@@ -136,8 +190,8 @@ public class SubmitLabel implements Runnable {
 				var p = points.get(j);
 				Element coordinate = doc.createElement("Coordinate");
 				coordinate.setAttribute("Order", String.valueOf(j));
-				coordinate.setAttribute("X", String.valueOf((int) p.getX()));
-				coordinate.setAttribute("Y", String.valueOf((int) p.getY()));
+				coordinate.setAttribute("X", String.valueOf((int) p.getX() - bbox[0]));
+				coordinate.setAttribute("Y", String.valueOf((int) p.getY() - bbox[1]));
 				coordinates.appendChild(coordinate);
 			}
 			count++;
@@ -185,5 +239,4 @@ public class SubmitLabel implements Runnable {
 			}
 		}
 	}
-
 }